+++ TinkoffASDKUI/TinkoffASDKUI/Modules/MainForm/Assembly/MainFormAssembly.swift
@@ -5,13 +5,33 @@
 //  Created by r.akhmadeev on 19.01.2023.
 //
 
+import TinkoffASDKCore
 import UIKit
 
 final class MainFormAssembly: IMainFormAssembly {
-    func build(stub: MainFormStub) -> UIViewController {
-        let router = MainFormRouter()
-        let presenter = MainFormPresenter(router: router, stub: stub)
+    private let coreSDK: AcquiringSdk
+
+    init(coreSDK: AcquiringSdk) {
+        self.coreSDK = coreSDK
+    }
+
+    func build(
+        paymentFlow: PaymentFlow,
+        configuration: MainFormUIConfiguration,
+        stub: MainFormStub
+    ) -> UIViewController {
+        let router = MainFormRouter(configuration: configuration)
+
+        let presenter = MainFormPresenter(
+            router: router,
+            coreSDK: coreSDK,
+            paymentFlow: paymentFlow,
+            configuration: configuration,
+            stub: stub
+        )
+
         let viewController = MainFormViewController(presenter: presenter)
+
         router.transitionHandler = viewController
         presenter.view = viewController