+++ TinkoffASDKUI/TinkoffASDKUI/Modules/MainForm/Router/MainFormRouter.swift
@@ -5,10 +5,22 @@
 //  Created by r.akhmadeev on 19.01.2023.
 //
 
+import TinkoffASDKCore
 import UIKit
 
 final class MainFormRouter: IMainFormRouter {
-    weak var transitionHandler: UIViewController?
+    // MARK: Dependencies
 
-    func openCardPaymentForm() {}
+    weak var transitionHandler: UIViewController?
+    private let configuration: MainFormUIConfiguration
+
+    // MARK: Init
+
+    init(configuration: MainFormUIConfiguration) {
+        self.configuration = configuration
+    }
+
+    // MARK: IMainFormRouter
+
+    func openCardPaymentForm(paymentFlow: PaymentFlow, cards: [PaymentCard]) {}
 }